requirements.txt
playwright==1.35.0
azure-keyvault-secrets==4.7.0
azure-identity==1.13.0
python-dotenv==1.0.0
cryptography==41.0.1
Office365-REST-Python-Client==2.4.1